// src/types/MovieTypes.ts
export interface IMovieTypes {
  backdrop_path: string
  first_air_date: string
  genre_ids: [number]
  id: number
  title: string
  name: string
  origin_country: [string]
  original_language: string
  original_name: string
  overview: string
  popularity: number
  poster_path: string
  vote_average: number
  vote_count: number
  media_type: string
  release_date: string
}